Defined various type names such as evnt_tid_t; Added a new function for counting the number of components in one event; Removed main.c and fixed Makefile, accordingly.

+++ tests/test_evnt_buffer_size.c
@@ -23,7 +23,7 @@ int main(int argc, const char **argv) {
     int i;
     long long int start, fin;
     const char* filename = "trace";
-    uint64_t buf_size;
+    uint32_t buf_size;
     evnt* event;
     trace buffer;
 
@@ -38,27 +38,27 @@ int main(int argc, const char **argv) {
     printf("What is the optimal buffer size for recording events?\n\n");
     printf("Buffer size (KB) \t Time \n");
 
-    buf_size = 1024;
+    buf_size = 1024; // 1KB
     while (buf_size <= MAX_BUFFER_SIZE) {
         init_trace(filename, buf_size);
 
         start = get_ticks();
         for (i = 0; i < (NB_EVENTS + 1) / 10; i++) {
-            evnt_probe0((uint64_t) 10 * i + 1);
-            evnt_probe1((uint64_t) 10 * i + 2, 1);
-            evnt_probe2((uint64_t) 10 * i + 3, 1, 3);
-            evnt_probe3((uint64_t) 10 * i + 4, 1, 3, 5);
-            evnt_probe4((uint64_t) 10 * i + 5, 1, 3, 5, 7);
-            evnt_probe5((uint64_t) 10 * i + 6, 1, 3, 5, 7, 11);
-            evnt_probe6((uint64_t) 10 * i + 7, 1, 3, 5, 7, 11, 13);
-            evnt_probe7((uint64_t) 10 * i + 8, 1, 3, 5, 7, 11, 13, 17);
-            evnt_probe8((uint64_t) 10 * i + 9, 1, 3, 5, 7, 11, 13, 17, 19);
-            evnt_probe9((uint64_t) 10 * i + 10, 1, 3, 5, 7, 11, 13, 17, 19, 23);
+            evnt_probe0(10 * i + 1);
+            evnt_probe1(10 * i + 2, 1);
+            evnt_probe2(10 * i + 3, 1, 3);
+            evnt_probe3(10 * i + 4, 1, 3, 5);
+            evnt_probe4(10 * i + 5, 1, 3, 5, 7);
+            evnt_probe5(10 * i + 6, 1, 3, 5, 7, 11);
+            evnt_probe6(10 * i + 7, 1, 3, 5, 7, 11, 13);
+            evnt_probe7(10 * i + 8, 1, 3, 5, 7, 11, 13, 17);
+            evnt_probe8(10 * i + 9, 1, 3, 5, 7, 11, 13, 17, 19);
+            evnt_probe9(10 * i + 10, 1, 3, 5, 7, 11, 13, 17, 19, 23);
         }
         fin = get_ticks();
 
         fin_trace();
-        printf("\t%lu\t\t %llu\n", buf_size / 1024, fin - start);
+        printf("\t%u\t\t %llu\n", buf_size / 1024, fin - start);
 
         buf_size = 2 * buf_size;
     }
@@ -70,7 +70,7 @@ int main(int argc, const char **argv) {
     printf("What is the optimal buffer size for reading the trace file?\n\n");
     printf("Buffer size (KB) \t Time \n");
 
-    buf_size = 1024;
+    buf_size = 1024; // 1KB
     while (buf_size <= MAX_BUFFER_SIZE) {
         set_read_buffer_size(buf_size);
         buffer = open_trace(filename);
@@ -85,7 +85,7 @@ int main(int argc, const char **argv) {
         fin = get_ticks();
 
         close_trace(&buffer);
-        printf("\t%lu\t\t %llu\n", buf_size / 1024, fin - start);
+        printf("\t%u\t\t %llu\n", buf_size / 1024, fin - start);
 
         buf_size = 2 * buf_size;
     }
